fix!: allow multi repos to load macros with same module name closes #3734
previously, loaders would load multiple projects at once, this is
problematic because macros with the same module name
(macros/__init__.py) would collide and not be picked up. this change
makes it so that we load each project sequentially. this also simplifies
the context because now we just have a list of projects instead of a
dictionary keyed by project type

this also addresses a bug where hybrid dbt/sqlmesh projects weren't
properly updating the DAG.